Trait image_processor::traits::Image

pub trait Image {
    fn dimensions(&self) -> &Dimensions;

    fn pixels(&self) -> Box<[u8]>;

    fn crop(&self, x: u32, y: u32, dims: Dimensions) -> ImageSlice;

    fn blurred(&self, amount: u32) -> OwnedImage;

    fn flipped(&self, horiz: bool, vert: bool) -> OwnedImage;

    fn greyscale(&self) -> OwnedImage;
}

Immutable image operations.

Required methods

fn dimensions(&self) -> &Dimensions

This image's width and height.

fn pixels(&self) -> Box<[u8]>

This image's pixels.

fn crop(&self, x: u32, y: u32, dims: Dimensions) -> ImageSlice

Borrows a sub-image.

fn blurred(&self, amount: u32) -> OwnedImage

Returns a new image that is this one blurred.

fn flipped(&self, horiz: bool, vert: bool) -> OwnedImage

Returns a new image that is this one flipped.

fn greyscale(&self) -> OwnedImage

Return s new image that is this one in grayscale

Loading content...